utils/pubsub_reader.py
from google.cloud import pubsub_v1
import json

#*******
    #Replace with actual Pub/Sub pull logic using SubscriberClient.pull() or BigQuery sink.
#*******




def get_pubsub_data():
    # Simulated agent status from Pub/Sub
    return {
        "Kiln": {"status": "✅ Active", "last_message": "Temp spike - adjusted"},
        "Cooler": {"status": "✅ Active", "last_message": "Flow synced"},
        "Mix": {"status": "✅ Active", "last_message": "Ratio rebalanced"},
        "Maintenance": {"status": "⚠️ Alert", "last_message": "Fault ID 23"}
    }