requirements.txt
streamlit==1.30.0
plotly
google-cloud-pubsub
google-cloud-monitoring
